// src/interfaces/request.interface.tsx
// Generated by https://quicktype.io
// CMD + SHIFT + P - Find Copy JSON Extension
// select language and the must create the inferface

export interface RequestList {
    page:        number;
    per_page:    number;
    total:       number;
    total_pages: number;
    data:        User[];
    support:     Support;
}

export interface User {
    id:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar:     string;
}

export interface Support {
    url:  string;
    text: string;
}